+++ test/integration-tests/lib/sniffer_error.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, PartialEq)]
+pub enum SnifferError {
+    DownstreamClosed,
+    UpstreamClosed,
+}
